Reorganize file structure and start implementing progress bar.
