+++ maxodiff-cli/src/main/java/org/monarchinitiative/maxodiff/cli/cmd/BaseLiricalCommand.java
@@ -1,10 +1,6 @@
 package org.monarchinitiative.maxodiff.cli.cmd;
 
-import org.monarchinitiative.lirical.core.analysis.AnalysisOptions;
-import org.monarchinitiative.lirical.core.exception.LiricalException;
 import org.monarchinitiative.lirical.core.model.TranscriptDatabase;
-import org.monarchinitiative.maxodiff.core.lirical.*;
-import org.monarchinitiative.phenol.ontology.data.TermId;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import picocli.CommandLine;
