+++ maxodiff-cli/src/main/java/org/monarchinitiative/maxodiff/cli/cmd/NewBenchmarkCommand.java
@@ -13,6 +13,7 @@ import org.monarchinitiative.lirical.io.analysis.PhenopacketData;
 import org.monarchinitiative.maxodiff.config.MaxodiffDataResolver;
 import org.monarchinitiative.maxodiff.config.MaxodiffPropsConfiguration;
 import org.monarchinitiative.maxodiff.core.analysis.*;
+import org.monarchinitiative.maxodiff.core.analysis.refinement.*;
 import org.monarchinitiative.maxodiff.core.io.PhenopacketFileParser;
 import org.monarchinitiative.maxodiff.core.lirical.*;
 import org.monarchinitiative.maxodiff.core.model.DifferentialDiagnosis;
