+++ maxodiff-core/src/main/java/org/monarchinitiative/maxodiff/core/analysis/refinement/MaxodiffResultImpl.java
@@ -1,4 +1,7 @@
-package org.monarchinitiative.maxodiff.core.analysis;
+package org.monarchinitiative.maxodiff.core.analysis.refinement;
+
+import org.monarchinitiative.maxodiff.core.analysis.Frequencies;
+import org.monarchinitiative.maxodiff.core.analysis.MaxoTermScore;
 
 import java.util.List;
