+++ maxodiff-core/src/main/java/org/monarchinitiative/maxodiff/core/analysis/refinement/RefinementResults.java
@@ -1,4 +1,4 @@
-package org.monarchinitiative.maxodiff.core.analysis;
+package org.monarchinitiative.maxodiff.core.analysis.refinement;
 
 import com.fasterxml.jackson.annotation.JsonGetter;
