+++ maxodiff-core/src/main/java/org/monarchinitiative/maxodiff/core/model/ExcludedPhenotypes.java
@@ -1,7 +1,7 @@
-package org.monarchinitiative.maxodiff.core.analysis;
+package org.monarchinitiative.maxodiff.core.model;
 
 import org.monarchinitiative.maxodiff.core.SimpleTerm;
-import org.monarchinitiative.maxodiff.core.model.Sample;
+import org.monarchinitiative.maxodiff.core.analysis.MaxoHpoTermIdMaps;
 import org.monarchinitiative.phenol.ontology.data.TermId;
 
 import java.util.*;
