+++ maxodiff-core/src/test/java/org/monarchinitiative/maxodiff/core/analysis/MaxoHpoTermProbabilitiesTest.java
@@ -4,6 +4,8 @@ import org.junit.jupiter.api.Test;
 import org.monarchinitiative.maxodiff.core.SimpleTerm;
 import org.monarchinitiative.maxodiff.core.TestResources;
 import org.monarchinitiative.maxodiff.core.model.DifferentialDiagnosis;
+import org.monarchinitiative.maxodiff.core.model.DiseaseModelProbability;
+import org.monarchinitiative.maxodiff.core.model.MaxoHpoTermProbabilities;
 import org.monarchinitiative.maxodiff.core.model.Sample;
 import org.monarchinitiative.phenol.annotations.formats.hpo.HpoDiseases;
 import org.monarchinitiative.phenol.ontology.data.TermId;
