+++ maxodiff-html/src/main/java/org/monarchinitiative/maxodiff/html/config/MaxodiffAutoConfiguration.java
@@ -4,8 +4,8 @@ import org.monarchinitiative.maxodiff.core.diffdg.DifferentialDiagnosisEngine;
 import org.monarchinitiative.maxodiff.core.io.MaxoDxAnnots;
 import org.monarchinitiative.maxodiff.config.MaxodiffDataException;
 import org.monarchinitiative.maxodiff.core.SimpleTerm;
-import org.monarchinitiative.maxodiff.core.analysis.DiffDiagRefiner;
-import org.monarchinitiative.maxodiff.core.analysis.MaxoDiffRefiner;
+import org.monarchinitiative.maxodiff.core.analysis.refinement.DiffDiagRefiner;
+import org.monarchinitiative.maxodiff.core.analysis.refinement.MaxoDiffRefiner;
 import org.monarchinitiative.maxodiff.config.MaxodiffDataResolver;
 import org.monarchinitiative.maxodiff.core.service.BiometadataService;
 import org.monarchinitiative.maxodiff.core.service.BiometadataServiceImpl;
